f add DockerFlow callbacks, fetch version from Cargo, add tests
