+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 #members = ["linkserver", "chatserver", "spake2_demo"]
-members = ["linkserver", "chatserver"]
+members = ["linkserver", "channelserver"]
