+++ channelserver/version.json
@@ -1,0 +1,6 @@
+{
+    "build": "TBD",
+    "commit": "TBD",
+    "source": "https://github.com/mozilla-services/pairsona",
+    "version": "TBD"
+}
